requirements.txt
# GlycoCast Diabetes Risk Calculator
# Requirements for V2

streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
fpdf2>=2.7.0
